Text script() was way too slow. Rewrote it in Java, and got a nice performance kick out of it, hopefully without causing too much damage to the codebase ;)
The generate_UnicodeScript.rb code only really needs to be run once, and I've already done that. I'll leave it in just in case we find some bugs in the future and need to fix it by changing the way the generated code output look (which I might add, is not very nice).

Signed-off-by: Ola Bini <ola.bini@gmail.com>